[Refactor]: weekCalendar 리팩토링 50%
