+++ src/components/common/calendar/datePicker/weekDatePicker.tsx
@@ -1,5 +1,8 @@
 'use client'
 import { useEffect, useState } from 'react'
+import { useRecoilValue, useSetRecoilState } from 'recoil'
+
+import { WeekCalendarDateState } from '@/lib/state/calendar/WeekCalendarDateState'
 
 import AllowLeftIcon from 'public/assets/icons/allow_left.svg'
 import AllowRightIcon from 'public/assets/icons/allow_right.svg'
@@ -10,18 +13,24 @@ interface dateType {
   date: number
 }
 
-export default function WeekDatePicker(props: any) {
+interface IProps {
+  handleChangeIsClickedDatePicker: () => void
+}
+
+export default function WeekDatePicker(props: IProps) {
+  const weekData = useRecoilValue(WeekCalendarDateState)
+  const setWeekData = useSetRecoilState(WeekCalendarDateState)
   const dateName = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
-  const [dateData, setDateData] = useState<dateType>(props.parentsDateData)
-  const [week, setWeek] = useState<number>(props.parentsWeekData)
-  const [clickedDate, setClickedDate] = useState<string>(dateData.date.toString())
-  const [isClickedDate, setIsClickedDate] = useState<boolean>(false)
+  //const [dateData, setDateData] = useState<dateType>(props.parentsDateData)
+  //const [week, setWeek] = useState<number>(props.parentsWeekData)
+  //const [clickedDate, setClickedDate] = useState<string>(weekData.date.toString())
+  //const [isClickedDate, setIsClickedDate] = useState<boolean>(false)
 
   const getCalanderData = () => {
-    const lastDateOfLastMonthData = new Date(dateData.year, dateData.month, 0)
-    const firstDateOfCurrentMonthData = new Date(dateData.year, dateData.month)
-    const lastDateOfCurrnetMonthData = new Date(dateData.year, dateData.month + 1, 0)
-    const firstDateOfNextMonthData = new Date(dateData.year, dateData.month + 1)
+    const lastDateOfLastMonthData = new Date(weekData.year, weekData.month, 0)
+    const firstDateOfCurrentMonthData = new Date(weekData.year, weekData.month)
+    const lastDateOfCurrnetMonthData = new Date(weekData.year, weekData.month + 1, 0)
+    const firstDateOfNextMonthData = new Date(weekData.year, weekData.month + 1)
 
     let list = []
     if (firstDateOfCurrentMonthData.getDay() !== 0) {
@@ -58,70 +67,53 @@ export default function WeekDatePicker(props: any) {
     return result
   }
 
-  const onClickDateHandler = (e: React.MouseEvent<HTMLDivElement>) => {
-    setWeek(Number(e.currentTarget.title))
-    setClickedDate(e.currentTarget.id)
-    setIsClickedDate(true)
+  const onClickDateHandler = (week: number) => {
+    setWeekData(prev => ({
+      ...prev,
+      week: week
+    }))
+
   }
 
   const onClickMonthForwardHandler = () => {
-    if (dateData.month === 11) {
-      setDateData((preDateData: dateType) => ({
-        ...preDateData,
-        year: preDateData.year + 1,
+    if (weekData.month === 11) {
+      setWeekData(prev => ({
+        ...prev,
+        year: prev.year + 1,
         month: 0
       }))
     } else {
-      setDateData((preDateData: dateType) => ({
-        ...preDateData,
-        month: preDateData.month + 1
+      setWeekData(prev => ({
+        ...prev,
+        month: prev.month + 1
       }))
     }
   }
 
   const onClickMonthBackHandler = () => {
-    if (dateData.month === 0) {
-      setDateData((preDateData: dateType) => ({
-        ...preDateData,
-        year: preDateData.year - 1,
+    if (weekData.month === 0) {
+      setWeekData(prev => ({
+        ...prev,
+        year: prev.year - 1,
         month: 11
       }))
     } else {
-      setDateData((preDateData: dateType) => ({
-        ...preDateData,
-        month: preDateData.month - 1
+      setWeekData(prev => ({
+        ...prev,
+        month: prev.month - 1
       }))
     }
   }
 
   const onClickCancelBtn = () => {
-    setClickedDate('')
-    setWeek(0)
+    props.handleChangeIsClickedDatePicker()
   }
 
   const onClickCheckHandler = () => {
-    props.changeParentsDateData({
-      year: dateData.year,
-      month: dateData.month,
-      date: Number(clickedDate)
-    })
-    props.changeParentsWeekData(week)
+    props.handleChangeIsClickedDatePicker()
   }
 
   const dateList = getCalanderData()
-
-  useEffect(() => {
-    if (isClickedDate) {
-      const dateList = getCalanderData()
-      for (var i = 0; i < dateList.length; i++) {
-        for (var j = 0; j < dateList[i].date.length; j++) {
-          if (dateList[i].date[j].date === clickedDate && dateList[i].date[j].clickable) {
-            setWeek(dateList[i].week)
-          }
-        }
-      }
-    }
-  }, [clickedDate])
 
   return (
     <div className="flex flex-col z-10 bg-white gap-2 w-[283px] p-4 rounded-lg shadow-[0px_1px_2px_-1px_rgba(0, 0, 0, 0.10)] shadow">
@@ -134,7 +126,7 @@ export default function WeekDatePicker(props: any) {
           onClick={onClickMonthBackHandler}
         />
         <div className="w-[126px] text-center gray-900-bold text-xs ">
-          {dateData.year}년 {dateData.month + 1}월
+          {weekData.year}년 {weekData.month + 1}월
         </div>
         <AllowRightIcon
           width={20}
@@ -163,15 +155,14 @@ export default function WeekDatePicker(props: any) {
                   <div
                     key={i}
                     id={dateData.date}
-                    title={data.week.toString()}
                     className={`px-1 py-2 cursor-pointer ${dateData.textColor} ${
-                      data.week === week && i !== 0 && i !== 6 ? 'bg-primary-50' : ''
-                    } ${data.week === week && i === 0 ? 'bg-primary-700 rounded-l-lg text-white font-bold' : ''} 
+                      data.week === weekData.week && i !== 0 && i !== 6 ? 'bg-primary-50' : ''
+                    } ${data.week === weekData.week && i === 0 ? 'bg-primary-700 rounded-l-lg text-white font-bold' : ''} 
                         ${
-                          data.week === week && i === 6 ? 'bg-primary-700 rounded-r-lg text-white font-bold' : ''
+                          data.week === weekData.week && i === 6 ? 'bg-primary-700 rounded-r-lg text-white font-bold' : ''
                         } text-xs text-center`}
-                    onClick={e => {
-                      dateData.clickable && onClickDateHandler(e)
+                    onClick={() => {
+                      dateData.clickable && onClickDateHandler(data.week)
                     }}
                   >
                     {dateData.date}
